revert: "revert: re-revert profile perk detail page w verify wallet and usernme (#2343)"
This reverts commit 72c2c6f0b5c7b6630be79dd7bcb3e8733e2b96a7.
