+++ src/components/Cards/PerksCard/PerksCard.tsx
@@ -21,7 +21,6 @@ interface PerksCardProps {
   isLoading?: boolean;
   isDisabled?: boolean;
   fullWidth?: boolean;
-  onClick?: () => void;
 }
 
 export const PerksCard = ({
@@ -33,7 +32,6 @@ export const PerksCard = ({
   isLoading,
   isDisabled = false,
   fullWidth,
-  onClick,
 }: PerksCardProps) => {
   if (isLoading) {
     return <PerksCardSkeleton fullWidth={fullWidth} />;
@@ -48,7 +46,6 @@ export const PerksCard = ({
         width: '100%',
         maxWidth: fullWidth ? '100%' : PERK_CARD_SIZES.CARD_WIDTH,
       }}
-      onClick={onClick}
     >
       <PerksCardActionArea disableRipple disabled={isDisabled}>
         {imageUrl ? (
