+++ src/components/Form/TaskInput/TaskInput.styles.ts
@@ -1,6 +1,5 @@
 import InputBase from '@mui/material/InputBase';
 import { styled } from '@mui/material/styles';
-import Typography from '@mui/material/Typography';
 
 export const TaskInputField = styled(InputBase)(({ theme }) => ({
   width: '100%',
@@ -32,12 +31,4 @@ export const TaskInputField = styled(InputBase)(({ theme }) => ({
       color: (theme.vars || theme).palette.textDisabled,
     },
   },
-  '&.Mui-error input': {
-    borderColor: (theme.vars || theme).palette.borderError,
-  },
-}));
-
-export const TaskInputErrorMessage = styled(Typography)(({ theme }) => ({
-  marginTop: theme.spacing(1),
-  color: (theme.vars || theme).palette.textError,
 }));
