--- src/components/ProfilePage/components/ClaimPerkModal/constants.ts
@@ -1,1 +1,0 @@
-export const MODAL_CONTAINER_ID = 'claim-perk-modal';
