--- src/stores/perkClaimStatus/index.ts
@@ -1,4 +1,0 @@
-export {
-  usePerkClaimStatusStore,
-  PerkClaimStatus,
-} from './PerkClaimStatusStore';
